requirements.txt
torch
torchvision
timm
numpy
tqdm    
matplotlib 
Pillow     